+++ pyproject.toml
@@ -4,6 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [tool.isort]
 profile = "black"
+line_length = 120
 
 [tool.black]
 line-length = 120
