+++ src/pyfreeleh/kv/__init__.py
@@ -1,6 +1,5 @@
 from typing import List
 
 from .gsheet import GoogleSheetKV
-from .base import KeyNotFoundError, KV
 
 __all__: List[str] = ["GoogleSheetKV"]
